requirements/ml.txt
# Machine Learning specific requirements
-r base.txt

# Advanced ML Libraries
prophet==1.1.5
mlflow==2.10.0
optuna==3.5.0
scipy==1.11.4
statsmodels==0.14.1

# Deep Learning (optional - uncomment if needed)
# torch==2.1.2
# tensorflow==2.15.0
# transformers==4.36.2

# Time Series Analysis
pandas-ta==0.3.14b0
influxdb-client==1.39.0

# Feature Engineering
feature-engine==1.6.2
category-encoders==2.6.3

# Model Interpretability
shap==0.44.0
lime==0.2.0.1

# AutoML (optional)
# auto-sklearn==0.15.0
# h2o==3.44.0.3

# Experiment Tracking
wandb==0.16.2
tensorboard==2.15.1

# Data Validation
great-expectations==0.18.8
evidently==0.4.13

# Feature Store
feast==0.36.0

# Model Serving
bentoml==1.1.11
ray[serve]==2.9.1